+++ helper-scripts/channel-offline.sh
@@ -45,7 +45,7 @@ xmltvloop=$(head -n 1 $workdir/xmlfiles4.txt)
 while [[ ! -z $xmltvloop ]]; do
 
 #randomise audio
-randomNumberoffline=$audionumber
+randomNumberoffline=$(shuf -i 1-$audioamount -n 1 --repeat)
 offlineaudio=$(head -n $randomNumberoffline $workdir/music.txt | tail -n 1)
 
 # get and read xmltv data
@@ -75,10 +75,6 @@ fi
 #news text colour
 if [[ $offlinetextcolour == random ]]
 then
-
-  randomise() {
-      shuf -i 1-$audioamount -n 1 --repeat
-  }
 
 #awk 'BEGIN{srand()}{print rand(), $0}' $workdir/colours.txt | sort -n -k 1 | awk 'sub(/\S* /,"")'
 #newstextcolourrandomNumber=$(shuf -i 1-140 -n 1 --repeat)