+++ helper-scripts/news.sh
@@ -13,17 +13,12 @@ fi
 processnews1=$(echo $processnews | tr '[:upper:]' '[:lower:]')
 if [[ $processnews1 = yes ]]
 then
-
-  randomise() {
-      shuf -i 1-$audioamount -n 1 --repeat
-  }
-
 #news
 
 #audio
-randomNumber3=$audionumber
-randomNumber4=$audionumber
-randomNumber5=$audionumber
+randomNumber3=$(shuf -i 1-$audioamount -n 1 --repeat)
+randomNumber4=$(shuf -i 1-$audioamount -n 1 --repeat)
+randomNumber5=$(shuf -i 1-$audioamount -n 1 --repeat)
 audio3=$(head -n $randomNumber3 $workdir/music.txt | tail -n 1)
 audio4=$(head -n $randomNumber4 $workdir/music.txt | tail -n 1)
 audio5=$(head -n $randomNumber5 $workdir/music.txt | tail -n 1)